requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
python-dotenv==0.14.0
python-kucoin==2.1.2
requests==2.24.0
urllib3==1.25.9
websockets==8.1
